MODIFY THE APP.RB FILE
