+++ app.rb
@@ -31,31 +31,5 @@ class App
     @teachers.each { |s| print "[#{s.class}] Name: #{s.name}, ID: #{s.id}, Age: #{s.age}\n" }
   end
 
-  
 
-  def create_book
-    print "Create book \n"
-    print ' Enter book title: '
-    title = gets.chomp
-    print ' Enter author name: '
-    author = gets.chomp
-    new_book = Book.new(title, author)
-    @books << new_book
-    puts " '#{new_book.title}' added successfully"
-  end
-
-  def create_rental
-    puts 'Enter date'
-    date = gets.chomp
-    puts 'Select book number'
-    list_books
-    book = gets.chomp.to_i
-    puts 'Enter person by number NOT ID'
-    people = [*@students, *@teachers]
-    people.each_with_index { |s, i| print "(#{i}) [#{s.class}] Name: #{s.name}, ID: #{s.id}, Age: #{s.age}\n" }
-    person = gets.chomp.to_i
-    new_rental = Rental.new(date, @books[book], people[person])
-    @rentals << new_rental
-    puts 'Rental added successfully'
-  end
 end
